backend/requirements.txt
python-dotenv==1.0.0
flask==3.0.0
flask-pydantic==0.12.0
prisma==0.11.0
flask-login==0.6.3
pyjwt==2.8.0
msal==1.28.0
retry==0.9.2
flask-cors==4.0.0
gunicorn==21.2.0
